# sha2: Add aarch64 backends for SHA2.