--- a/sha2/src/sha512.rs
+++ b/sha2/src/sha512.rs
@@ -15,7 +15,7 @@ cfg_if::cfg_if! {
         }
         mod x86;
         use x86::compress;
-    } else if #[cfg(target_arch = "aarch64")] {
+    } else if #[cfg(all(feature = "asm", target_arch = "aarch64"))] {
         mod soft;
         mod aarch64;
         use aarch64::compress;