--- a/sha2/src/sha512/aarch64.rs
+++ b/sha2/src/sha512/aarch64.rs
@@ -188,7 +188,11 @@ unsafe fn vsha512hq_u64(
     hash_gf: uint64x2_t,
     kwh_kwh2: uint64x2_t,
 ) -> uint64x2_t {
-    asm!("SHA512H {:q}, {:q}, {:v}.2D", inout(vreg) hash_ed, in(vreg) hash_gf, in(vreg) kwh_kwh2);
+    asm!(
+        "SHA512H {:q}, {:q}, {:v}.2D",
+        inout(vreg) hash_ed, in(vreg) hash_gf, in(vreg) kwh_kwh2,
+        options(pure, nomem, nostack, preserves_flags)
+    );
     hash_ed
 }
 
@@ -198,13 +202,21 @@ unsafe fn vsha512h2q_u64(
     hash_c_: uint64x2_t,
     hash_ab: uint64x2_t,
 ) -> uint64x2_t {
-    asm!("SHA512H2 {:q}, {:q}, {:v}.2D", inout(vreg) sum_ab, in(vreg) hash_c_, in(vreg) hash_ab);
+    asm!(
+        "SHA512H2 {:q}, {:q}, {:v}.2D",
+        inout(vreg) sum_ab, in(vreg) hash_c_, in(vreg) hash_ab,
+        options(pure, nomem, nostack, preserves_flags)
+    );
     sum_ab
 }
 
 #[inline(always)]
 unsafe fn vsha512su0q_u64(mut w0_1: uint64x2_t, w2_: uint64x2_t) -> uint64x2_t {
-    asm!("SHA512SU0 {:v}.2D, {:v}.2D", inout(vreg) w0_1, in(vreg) w2_);
+    asm!(
+        "SHA512SU0 {:v}.2D, {:v}.2D",
+        inout(vreg) w0_1, in(vreg) w2_,
+        options(pure, nomem, nostack, preserves_flags)
+    );
     w0_1
 }
 
@@ -214,6 +226,10 @@ unsafe fn vsha512su1q_u64(
     w14_15: uint64x2_t,
     w9_10: uint64x2_t,
 ) -> uint64x2_t {
-    asm!("SHA512SU1 {:v}.2D, {:v}.2D, {:v}.2D", inout(vreg) s01_s02, in(vreg) w14_15, in(vreg) w9_10);
+    asm!(
+        "SHA512SU1 {:v}.2D, {:v}.2D, {:v}.2D",
+        inout(vreg) s01_s02, in(vreg) w14_15, in(vreg) w9_10,
+        options(pure, nomem, nostack, preserves_flags)
+    );
     s01_s02
 }