@tarcieri 2023-06-15
Neat! I hadn't thought about using inline ASM as a sort of "polyfill" for using unstable intrinsics on stable Rust before.

This is something we should consider doing elsewhere we use unstable aarch64 intrinsics, such as in the `aes` and `polyval` crates. Then, when the intrinsics are stabilized, we can delete the inline ASM and switch to the intrinsics.