@newpavlov 2023-06-15
Update the cross version to fix the CI failures.