@codahale 2023-06-15
Now that this is merged, what’s the remaining work to drop the `sha2_asm` dependency?